feat: Refactor as a monorepo, using Lerna
