+++ .gitignore
@@ -1,37 +1,3 @@
+.*
 /example/i18n.js
-/messageformat.*js*
 node_modules/
-
-# Numerous always-ignore extensions
-*.diff
-*.err
-*.orig
-*.log
-*.rej
-*.swo
-*.swp
-*.vi
-*~
-*.sass-cache
-
-# OS or Editor files
-.DS_Store
-Thumbs.db
-.cache
-.project
-.settings
-.tmproj
-*.esproj
-nbproject
-*.sublime-project
-*.sublime-workspace
-.vimrc
-
-# Komodo
-*.komodoproject
-.komodotools
-
-# Folders to ignore
-intermediate
-publish
-.idea
