+++ packages/cli/package.json
@@ -21,14 +21,14 @@
   "engines": {
     "node": ">=6.0"
   },
-  "scripts": {
-    "prepare": "npm install --no-save ../"
-  },
   "dependencies": {
     "dot-properties": "^0.2.1",
     "glob": "~7.0.6",
     "nopt": "~3.0.6",
     "uv": "^1.0.7"
+  },
+  "devDependencies": {
+    "messageformat": "^2.0.4"
   },
   "peerDependencies": {
     "messageformat": "2.x"
