+++ packages/messageformat/lib/compiler.js
