+++ packages/messageformat/lib/formatters/duration.js
