+++ packages/messageformat/lib/formatters/number.js
