+++ packages/messageformat/lib/formatters/time.js
