+++ packages/messageformat/lib/messageformat.js
