+++ packages/messageformat/lib/plurals.js
