+++ packages/messageformat/package.json
@@ -1,0 +1,50 @@
+{
+  "name": "messageformat",
+  "version": "2.0.5",
+  "description": "PluralFormat and SelectFormat Message and i18n Tool - A JavaScript Implemenation of the ICU standards.",
+  "keywords": [
+    "i18n",
+    "pluralformat",
+    "icu",
+    "gettext",
+    "choiceformat",
+    "selectformat",
+    "messageformat",
+    "internationalization"
+  ],
+  "author": "Alex Sexton <alexsexton@gmail.com>",
+  "contributors": [
+    "Eemeli Aro <eemeli@gmail.com>"
+  ],
+  "license": "MIT",
+  "homepage": "https://messageformat.github.io/messageformat/",
+  "repository": "messageformat/messageformat",
+  "main": "lib/messageformat.js",
+  "files": [
+    "example",
+    "index.d.ts",
+    "lib",
+    "messageformat.js",
+    "messageformat.min.js",
+    "messageformat.min.js.map",
+    "messages.js"
+  ],
+  "dependencies": {
+    "make-plural": "^4.3.0",
+    "messageformat-parser": "^3.0.0",
+    "reserved-words": "^0.1.2"
+  },
+  "devDependencies": {
+    "browserify": "^16.2.3",
+    "uglify-js": "^3.4.9"
+  },
+  "scripts": {
+    "build:browser": "browserify lib/messageformat.js -s MessageFormat -o messageformat.js",
+    "build:minify": "uglifyjs messageformat.js --compress --mangle --output messageformat.min.js --source-map \"url='messageformat.min.js.map'\"",
+    "build": "npm run build:browser && npm run build:minify",
+    "clean": "rimraf messageformat.*js*"
+  },
+  "browser": {
+    "fs": false
+  }
+}
