+++ test/index.html
@@ -7,7 +7,7 @@
     <script src="../node_modules/mocha/mocha.js"></script>
     <script src="../node_modules/expect.js/index.js"></script>
     <script>mocha.setup('bdd')</script>
-    <script src="../messageformat.js"></script>
+    <script src="../packages/messageformat/messageformat.js"></script>
     <script src="messageformat.js"></script>
   </head>
   <body>
